fix: rename utils to serialize
